pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ids-agent"
version = "2.0.0"
description = "IDS Agent for Raspberry Pi - Security monitoring with Suricata"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SIXT R&D Team", email = "dev@sixt.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
]
keywords = ["ids", "security", "suricata", "raspberry-pi", "monitoring"]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "punq>=0.6",
    "aiohttp>=3.8",
    "redis>=4.5",
    "boto3>=1.26",
    "requests>=2.28",
    "requests-aws4auth>=1.2",
    "opensearch-py>=2.4.0",
    "toml>=0.10",
    "prometheus-client>=0.17",
    "python-json-logger>=2.0",
    "dataclasses-json>=0.5.14",
    "fastapi>=0.110.0",
    "uvicorn>=0.23.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "pytest-mock>=3.10",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "isort>=5.12",
    "bandit>=1.7",
    "pylint>=3.0",
    "safety>=2.3",
]

[project.scripts]
ids-agent = "ids.app.supervisor:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120
target-version = ['py310', 'py311']

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers -v"
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "unit: Unit tests (isolated)",
    "integration: Integration tests",
    "aws: Tests requiring AWS connectivity",
    "state_machine: Tests for state machine transitions",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/ids"]
branch = true

# ============================================================================
# RUFF - Fast Python Linter (Rust-based)
# ============================================================================
[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by black)
    "PLR0913", # too many arguments
    "PLR0911", # too many return statements
    "PLR0912", # too many branches
    "PLR2004", # magic value comparison
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004"]
"__init__.py" = ["F401"]  # unused imports in __init__

[tool.ruff.lint.isort]
known-first-party = ["ids"]

# ============================================================================
# MYPY - Static Type Checker
# ============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
strict_equality = true
strict_concatenate = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true
pretty = true

# Strictness levels per module
[[tool.mypy.overrides]]
module = "ids.tailscale.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "ids.domain.*"
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "pyvis.*",
    "tailscale.*",
    "punq.*",
    "opensearchpy.*",
    "redis.*",
    "prometheus_client.*",
]
ignore_missing_imports = true

# ============================================================================
# RADON - Code Complexity Analysis
# ============================================================================
# Run with: radon cc src/ids -a -s
# Run with: radon mi src/ids -s
# 
# Cyclomatic Complexity grades:
#   A (1-5)   - Low risk
#   B (6-10)  - Low risk  
#   C (11-20) - Moderate risk
#   D (21-30) - High risk
#   E (31-40) - Very high risk
#   F (41+)   - Untestable
#
# Maintainability Index:
#   A (20-100) - Very maintainable
#   B (10-19)  - Maintainable
#   C (0-9)    - Difficult to maintain

# ============================================================================
# BANDIT - Security Linter
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # assert_used - OK in tests
